Restructure into installable package
